requirements.txt
ollama>=0.4.0
pyaudio==0.2.13
pydub==0.25.1
python-dotenv==1.0.1
SpeechRecognition==3.10.0
gTTS==2.5.0
simpleaudio==1.0.4
pocketsphinx==5.0.4 